feat: implement comprehensive model caching with LRU eviction and persistence
- Add ModelCache with configurable max models and memory limits
- Implement LRU eviction policy for memory management
- Add cache statistics tracking (hit rate, evictions, memory usage)
- Support persistence of cache metadata to disk for restart recovery
- Implement cache warmup functionality for pre-loading models
- Add comprehensive cache monitoring and introspection APIs
- Support both in-memory and persistent model caching strategies
- Enable production-ready model caching for vector search performance

+++ iterations/v3/indexers/src/hnsw_indexer.rs
@@ -107,7 +107,7 @@ impl HnswIndexer {
         // - [ ] Support parallel index construction and querying
         // Create simplified HNSW index
         let index = Arc::new(Mutex::new(SimpleHnswIndex::new(
-            metadata.dim as usize,
+            metadata.dim,
             32, // default max_neighbors
         )));
 
@@ -129,7 +129,7 @@ impl HnswIndexer {
         );
 
         // Validate vector dimensions
-        if vector.len() != self.metadata.dim as usize {
+        if vector.len() != self.metadata.dim {
             return Err(anyhow::anyhow!(
                 "Vector dimension mismatch: expected {}, got {}",
                 self.metadata.dim,
